Dont let create board and task if the title or name exists

+++ src/main/java/com/api/kanban_board/services/tasks/GetTaskByIdService.java
@@ -13,7 +13,7 @@ public class GetTaskByIdService {
         this.taskRepository = taskRepository;
     }
 
-    public TaskModel execute(Long id){
+    public TaskModel execute(Integer id){
         return taskRepository.getTaskById(id);
     }
 }
